pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiwriter"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
    "httpx>=0.27.0",
    "trafilatura>=1.12.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "python-frontmatter>=1.1.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
]

[project.scripts]
aiwriter = "aiwriter.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["aiwriter"]
